fix, feat, refactor: 타입 분리 작업 마무리(불필요 import 제거), 데덴찌 메뉴 내 버튼 글씨 색 검정으로 변경, 뽑기 다음 카드 뽑으면 이전카드 바로 사라지도록 설정, 섞을때 이전 카드 바로 사라지도록 설정, 자동 팀분배 로직 파일 이름 변경
