+++ src/app/rock-scissor-paper/result/RspRoom.tsx
@@ -9,8 +9,8 @@ import { useEffect, useRef, useState } from 'react';
 import Play from '@/components/Play';
 import RspOptions from './RspOptions';
 import Button from '@/components/Button';
-import { rspResponse, toResult } from '@/utils/rspResultConverter';
-import RspResult, { RspResultType } from './RspResult';
+import { toResult } from '@/utils/rspResultConverter';
+import RspResult from './RspResult';
 
 export default function RspRoom() {
   const searchParams = useSearchParams();
