+++ src/utils/autoSplitTeams.ts
